Fixed some more issues

+++ thermos/start.sh
@@ -11,12 +11,15 @@ if [ -f $CHK_FILE ]; then
 else
    echo "Downloading required filed"
    wget https://github.com/CyberdyneCC/Thermos/releases/download/58/Thermos-1.7.10-1614-server.jar -O ${SERVER_JARFILE}
+fi
+
+cd /home/container
+
+if [ ! -d "libaries" ]; then
    wget https://github.com/CyberdyneCC/Thermos/releases/download/58/libraries.zip
    unzip libraries.zip
    rm libraries.zip
 fi
-
-cd /home/container
 
 if [ -z "$STARTUP"  ]; then
     # Output java version to console for debugging purposes if needed.
